+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 
 - Radial filter: Source => TNR and vice versa
+- Taxonomic tree: Options to show/hide descendants count and Level-based coloring.
 
 ### Fixed